Break out breadcrumbs and search result into their own style files
