+++ components/resources/ResourceSearchResult.js
@@ -20,7 +20,7 @@ import Button from "components/Button";
 import { useRouter } from "next/router";
 import LabelWithValue from "components/LabelWithValue";
 import { useTheme } from "providers/theme";
-import styles from "styles/modules/Search.module.scss";
+import styles from "styles/modules/SearchResult.module.scss";
 import { showError } from "utils/toast-utils";
 
 const ResourceSearchResult = ({ searchResult }) => {
