+++ styles/modules/Breadcrumbs.module.scss
@@ -1,0 +1,69 @@
+/**
+ * Copyright 2021 The Rode Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+@import "styles/constants";
+@import "styles/mixins";
+
+.breadcrumbs {
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+  justify-content: flex-start;
+  align-items: center;
+  padding: 0.5rem 1rem;
+
+  > * {
+    padding: 0 0.25rem;
+  }
+
+  > a {
+    @include ellipsis(calc(100vw - 12rem));
+  }
+
+  @include tabletAndLarger {
+    > a {
+      max-width: calc(100vw - 22rem);
+    }
+  }
+}
+
+.rootCrumb {
+  display: none;
+
+  @include tabletAndLarger {
+    display: unset;
+  }
+}
+
+.darkTheme {
+  &.breadcrumbs {
+    color: $DT_SUB_TEXT;
+
+    a {
+      color: $DT_LINK;
+    }
+  }
+}
+
+.lightTheme {
+  &.breadcrumbs {
+    color: $LT_SUB_TEXT;
+
+    a {
+      color: $LT_LINK;
+    }
+  }
+}
